feat: split core.js movieApi
