+++ service/movieApi/src/core/_startGenerateImageAndMovie.js
@@ -1,0 +1,1 @@
+
