+++ service/movieApi/src/core/handleFileList.js
@@ -1,0 +1,12 @@
+import { mod, store } from './init.js'
+export default {}
+
+export const handleFileList = ({ requestId }) => {
+  const MOVIE_DIR_PATH = mod.setting.getValue('path.MOVIE_DIR_PATH') 
+  const dirPath = requestId? `${MOVIE_DIR_PATH}${requestId}/`: MOVIE_DIR_PATH
+  console.log({ dirPath })
+  const fileDirList = mod.input.getFileDirListWithTitle({ dirPath })
+  const handleResult = { result: { fileDirList } }
+  return handleResult
+}
+
