+++ service/movieApi/src/core/handleRegisterDummyPrompt.js
@@ -1,0 +1,32 @@
+import { mod, store } from './init.js'
+export default {}
+
+export const handleRegisterDummyPrompt = async ({}) => {
+  const queue = mod.setting.getValue('amqp.REQUEST_QUEUE') 
+  await mod.amqpChannel.assertQueue(queue)
+
+  const requestId = mod.lib.getUlid()
+  const messageBuffer = _getDummyRequest({ requestId })
+  mod.amqpChannel.sendToQueue(queue, messageBuffer)
+
+  const handleResult = { isRegistered: true, requestId }
+  return handleResult
+}
+
+const _getDummyRequest = ({ requestId }) => {
+  const requestType = 'main_dummy'
+
+  const currentDelimiter = Buffer.from(mod.lib.getUlid())
+  const delimiterDelimiter = Buffer.from('|')
+  const messageBuffer = Buffer.concat([
+    currentDelimiter,
+    delimiterDelimiter,
+    Buffer.from(requestType),
+    currentDelimiter,
+    Buffer.from(requestId),
+    currentDelimiter,
+  ])
+
+  return messageBuffer
+}
+
